Initial attempt at fixing the "open surface" problem.
The cerebrum surface joining routine in Deformable fails if there are 
gaps in the "regions.nii.gz" label image at the midline.
A possible fix is to increase the structuring element size of the 
closings if it fails. If the intial size of 5 fails then try again with 
a larger structuring element 15.
If the larger structuring element fails then quit with an error message 
and don't do the freesurfer conversion step.